ENH: Add DisplacementFieldSubsamplingFactor
Output ANTs displacement fields are very dense, sampled with the fixed
image. This is generally over-sampled for the purpose of downstream use.
Memory usage is very high and serialization and deserialization takes
quite some time.

Add a DisplacementFieldSubsamplingFactor parameter for downsampling the
resulting displacement fields. This is applied in all directions for
both the forward and inverse transform. If the
DisplacementFieldSubsamplingFactor is greater than 1, this is applied.
The current default is 2.

This uses the itk::DisplacementFieldTransformParametersAdapter. In the
future, we may want to increase the default, and / or use the
itk::GaussianSmoothingOnUpdateDisplacementFieldTransformParametersAdaptor
with larger factors to avoid aliasing.